Refactored code base and directories to ensure PEP8 compliance
Optimised A* Algorithm

Converted repo into git package

A* exhibiting undefined behaviour

Potentially fixed A* bug

Minor bugfix

Identified x->y bug

changed grid indexing conventions

Squashed bugs by returning to master version of a star

Updated README
