+++ nova_rover_demos/pose_control.py
@@ -13,7 +13,7 @@ P. I. Corke, "Robotics, Vision & Control", Springer 2017, ISBN 978-3-319-54413-7
 import matplotlib.pyplot as plt
 import numpy as np
 from random import random
-from demoUtilities import rover
+from utils.rover import Rover
 # simulation parameters
 Kp_rho = 9
 Kp_alpha = 15
