+++ nova_rover_demos/utils/rover.py
@@ -1,4 +1,4 @@
-class rover:
+class Rover:
     def __init__(self,x,y,th):
         # Initial conditions selected arbitrarily
         self.x = x
@@ -6,10 +6,12 @@ class rover:
         self.theta = th
         self.v = 0
         self.world = []
-    def updatestate(self,state):
+        
+    def update_state(self,state):
         self.x = state[0]
         self.y = state[1]
         self.theta = state[2]
         self.v = state[3]
-    def updateMap(self,map):
+        
+    def update_map(self,map):
         self.world = map
